requirements.txt
absl-py==1.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
blis==0.7.9
Brlapi==0.8.4
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
command-not-found==0.3
confection==0.0.4
contourpy==1.0.6
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
dbus-python==1.2.18
debugpy==1.6.4
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.8.0
distro-info==1.1+ubuntu1
Django==4.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
duplicity==0.8.22
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
fasteners==0.18
fastjsonschema==2.16.2
flatbuffers==23.3.3
fonttools==4.38.0
fpdf==1.7.2
fqdn==1.5.1
future==0.18.2
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
gyp==0.1
h11==0.14.0
h5py==3.8.0
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
jaraco.classes==3.2.3
jax==0.4.8
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
kaleido==0.2.1
keras==2.12.0
keyring==23.11.0
kiwisolver==1.4.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.18
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
libclang==16.0.0
lockfile==0.12.2
louis==3.22.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
ml-dtypes==0.1.0
monotonic==1.6
more-itertools==9.0.0
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.11.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
opt-einsum==3.3.0
outcome==1.2.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pathy==0.10.1
pdfrw==0.4
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.3.0
pip-review==1.3.0
pipdeptree==2.7.0
platformdirs==2.6.0
plotly==5.11.0
plotly-gif==0.0.3
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.6
Pygments==2.13.0
PyGObject==3.42.2
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyPDF2==2.12.1
pyrsistent==0.19.2
python-apt==2.4.0+22.10
python-dateutil==2.8.2
python-debian==0.1.49
python-json-logger==2.0.4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.28
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.2
SecretStorage==3.3.3
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.6.2
systemd-python==235
tenacity==8.1.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
terminado==0.17.1
testpath==0.6.0
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toposort==1.7
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
usb-creator==0.3.7
wadllib==1.3.6
wasabi==1.1.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.4
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
xdg==5.1.1
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==3.0.3
yellowbrick==1.5
youtube-dl==2021.12.17
zipp==3.11.0
